+++ src/cmd/install/utils.go
@@ -2,96 +2,86 @@ package cmdlipinstall
 
 import (
 	"archive/zip"
-	"bufio"
 	"errors"
 	"io"
-	"net/http"
 	"os"
 	"os/exec"
 	"path/filepath"
-	"regexp"
 	"runtime"
-	"sort"
-	"strconv"
 	"strings"
 
 	"github.com/liteldev/lip/context"
 	"github.com/liteldev/lip/localfile"
+	"github.com/liteldev/lip/specifiers"
 	"github.com/liteldev/lip/tooth/toothfile"
 	"github.com/liteldev/lip/tooth/toothrecord"
 	"github.com/liteldev/lip/utils/download"
 	"github.com/liteldev/lip/utils/logger"
-	versionutils "github.com/liteldev/lip/utils/version"
 )
 
-// downloadTooth downloads a tooth file from a tooth repository, a tooth url,
-// or a local path and returns the path of the downloaded tooth file.
-// If the specifier is a requirement specifier, it should contain version.
-func downloadTooth(specifier Specifier) (string, error) {
-	switch specifier.Type() {
-	case ToothFileSpecifierType:
-		// For local tooth file, just return the path.
-
+// getTooth gets the tooth file path of a tooth specifier either from the cache or from the tooth repository.
+// If the tooth file is downloaded, it will be cached.
+// If the specifier is local tooth file, it will return the path of the local tooth file.
+// toothFilePath is the absolute path of the tooth file.
+func getTooth(specifier specifiers.Specifier) (isCached bool, toothFilePath string, err error) {
+	// For local tooth file, return the path directly.
+	if specifier.Type() == specifiers.ToothFileSpecifierType {
 		// Get full path of the tooth file.
 		toothFilePath, err := filepath.Abs(specifier.ToothFilePath())
 		if err != nil {
-			return "", errors.New("cannot get full path of tooth file: " + specifier.ToothFilePath())
+			return false, "", errors.New("cannot get full path of tooth file: " + specifier.ToothFilePath())
 		}
 
-		return toothFilePath, nil
+		return false, toothFilePath, nil
+	}
 
-	case ToothURLSpecifierType:
+	// Get the path to the cache tooth file.
+	cacheFileName := localfile.GetCachedToothFileName(specifier.String())
+	cacheDirectory, err := localfile.CacheDir()
+	if err != nil {
+		return false, "", err
+	}
+	cacheFilePath := filepath.Join(cacheDirectory, cacheFileName)
+
+	// Directly return the cached tooth file path if it exists.
+	isCacheExist, err := localfile.IsCachedToothFileExist(specifier.String())
+	if err != nil {
+		return false, "", err
+	}
+	if isCacheExist {
+		return true, cacheFilePath, nil
+	}
+
+	// Download the tooth file to the cache.
+	err = DownloadTooth(specifier, cacheFilePath)
+	if err != nil {
+		return false, "", err
+	}
+
+	return false, cacheFilePath, nil
+}
+
+// DownloadTooth downloads a tooth file from a tooth repository, a tooth url,
+// or a local path and returns the path of the downloaded tooth file.
+// If the specifier is a requirement specifier, it should contain version.
+func DownloadTooth(specifier specifiers.Specifier, destination string) error {
+	switch specifier.Type() {
+	case specifiers.ToothFileSpecifierType:
+		// Local tooth file is not accepted here.
+		return errors.New("local tooth file is not able to be downloaded")
+
+	case specifiers.ToothURLSpecifierType:
 		// For tooth url, download the tooth file and return the path.
 
-		cacheFileName := localfile.GetCachedToothFileName(specifier.String())
-
-		// Directly return the cached tooth file path if it exists.
-		isCacheExist, err := localfile.IsCachedToothFileExist(specifier.String())
+		err := download.DownloadFile(specifier.ToothURL(), destination)
 		if err != nil {
-			return "", err
+			return err
 		}
 
-		if isCacheExist {
-			cacheDir, err := localfile.CacheDir()
-			if err != nil {
-				return "", err
-			}
-			return cacheDir + "/" + cacheFileName, nil
-		}
+		return nil
 
-		// Download the tooth file to the cache.
-		cacheDir, err := localfile.CacheDir()
-		if err != nil {
-			return "", err
-		}
-
-		cacheFilePath := cacheDir + "/" + cacheFileName
-
-		err = download.DownloadFile(specifier.ToothURL(), cacheFilePath)
-		if err != nil {
-			return "", err
-		}
-
-		return cacheFilePath, nil
-
-	case RequirementSpecifierType:
+	case specifiers.RequirementSpecifierType:
 		// For requirement specifier, download the tooth via GOPROXY and return the path.
-
-		cacheFileName := localfile.GetCachedToothFileName(specifier.String())
-
-		// Directly return the cached tooth file path if it exists.
-		isCacheExist, err := localfile.IsCachedToothFileExist(specifier.String())
-		if err != nil {
-			return "", err
-		}
-
-		if isCacheExist {
-			cacheDir, err := localfile.CacheDir()
-			if err != nil {
-				return "", err
-			}
-			return cacheDir + "/" + cacheFileName, nil
-		}
 
 		// Get the tooth file url.
 		urlSuffix := "+incompatible.zip"
@@ -100,68 +90,16 @@ func downloadTooth(specifier Specifier) (string, error) {
 		}
 		url := context.Goproxy + "/" + specifier.ToothRepo() + "/@v/v" + specifier.ToothVersion().String() + urlSuffix
 
-		// Download the tooth file to the cache.
-		cacheDir, err := localfile.CacheDir()
+		err := download.DownloadFile(url, destination)
 		if err != nil {
-			return "", err
+			return err
 		}
 
-		cacheFilePath := cacheDir + "/" + cacheFileName
-
-		err = download.DownloadFile(url, cacheFilePath)
-		if err != nil {
-			return "", err
-		}
-
-		return cacheFilePath, nil
+		return nil
 	}
 
 	// Default to unknown error.
-	return "", errors.New("unknown error")
-}
-
-// FetchVersionList fetches the version list of a tooth repository.
-func FetchVersionList(repoPath string) ([]versionutils.Version, error) {
-	if !isValidRepoPath(repoPath) {
-		return nil, errors.New("invalid repository path: " + repoPath)
-	}
-
-	url := context.Goproxy + "/" + repoPath + "/@v/list"
-
-	// To lowercases.
-	url = strings.ToLower(url)
-
-	// Get the version list.
-	resp, err := http.Get(url)
-	if err != nil {
-		return nil, errors.New("cannot access GOPROXY: " + context.Goproxy)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		return nil, errors.New("cannot access tooth repository (HTTP CODE " + strconv.Itoa(resp.StatusCode) + "): " + repoPath)
-	}
-
-	// Each line is a version.
-	var versionList []versionutils.Version
-	scanner := bufio.NewScanner(resp.Body)
-	for scanner.Scan() {
-		versionString := scanner.Text()
-		versionString = strings.TrimPrefix(versionString, "v")
-		versionString = strings.TrimSuffix(versionString, "+incompatible")
-		version, err := versionutils.NewFromString(versionString)
-		if err != nil {
-			continue
-		}
-		versionList = append(versionList, version)
-	}
-
-	// Sort the version list in descending order.
-	sort.Slice(versionList, func(i, j int) bool {
-		return versionutils.GreaterThan(versionList[i], versionList[j])
-	})
-
-	return versionList, nil
+	return errors.New("unknown error")
 }
 
 // Install installs the .tth file.
@@ -313,46 +251,6 @@ func Install(t toothfile.ToothFile, isManuallyInstalled bool, isYes bool) error 
 	err = os.WriteFile(recordFilePath, recordJSON, 0755)
 	if err != nil {
 		return errors.New("failed to write record file " + recordFilePath + " " + err.Error())
-	}
-
-	return nil
-}
-
-// isValidRepoPath checks if the repoPath is valid.
-func isValidRepoPath(repoPath string) bool {
-	reg := regexp.MustCompile(`^[a-zA-Z\d-_\.\/]*$`)
-
-	// If not matched or the matched string is not the same as the specifier, it is an
-	// invalid requirement specifier.
-	return reg.FindString(repoPath) == repoPath
-}
-
-// validateToothRepoVersion checks if the version of the tooth repository is valid.
-func validateToothRepoVersion(repoPath string, version versionutils.Version) error {
-	if !isValidRepoPath(repoPath) {
-		return errors.New("invalid repository path: " + repoPath)
-	}
-
-	// Check if the version is valid.
-	urlSuffix := "+incompatible.info"
-	if strings.HasPrefix(version.String(), "0.") || strings.HasPrefix(version.String(), "1.") {
-		urlSuffix = ".info"
-	}
-	url := context.Goproxy + "/" + repoPath + "/@v/v" + version.String() + urlSuffix
-
-	// To lower case.
-	url = strings.ToLower(url)
-
-	// Get the version information.
-	resp, err := http.Get(url)
-	if err != nil {
-		return errors.New("cannot access GOPROXY: " + context.Goproxy)
-	}
-	defer resp.Body.Close()
-
-	// If the status code is 200, the version is valid.
-	if resp.StatusCode != 200 {
-		return errors.New("cannot access tooth (HTTP CODE " + strconv.Itoa(resp.StatusCode) + "): " + repoPath + "@" + version.String())
 	}
 
 	return nil